new folder structure + be fluent ai UI
