+++ src/app/exercises/modal-can/page.tsx
@@ -3,10 +3,10 @@
 import { useRouter } from 'next/navigation';
 import { motion } from 'framer-motion';
 import { useState } from 'react';
-import { useFeedback } from '../../components/game/FeedbackProvider';
-import { MobileSentenceBuilder, StepConfig, StepOption } from '../../components/game/MobileSentenceBuilder';
-import { ShareButton } from '../../components/game/ShareButton';
-import { AudioPlayer } from '../../components/game/AudioPlayer';
+import { useFeedback } from '../../../components/game/FeedbackProvider';
+import { MobileSentenceBuilder, StepConfig, StepOption } from '../../../components/game/MobileSentenceBuilder';
+import { ShareButton } from '../../../components/game/ShareButton';
+import { AudioPlayer } from '../../../components/game/AudioPlayer';
 
 // --- DATA DEFINITIONS FOR LEVEL 5 (EXPANDED & MORE USEFUL) ---
 const subjectOptions: StepOption[] = [
@@ -86,7 +86,7 @@ const Level5Page = () => {
 
   const handleNextLevel = () => {
     trackLevelCompletion(7);
-    router.push('/modal-could'); 
+    router.push('/exercises/modal-could'); 
   };
 
   const getSmartTranslation = () => {
